+++ src/Components/Pages/Visualizer/Graphs/Computation/SkeletonPbft.js
@@ -1,5 +1,5 @@
-import { ACTION_TYPE_PBFT_GRAPH, COLORS_PBFT_GRAPH, NODES_PBFT_GRAPH, TITLES_PBFT_GRAPH } from "../../../../../../Constants";
-import { dummyData } from "../../data";
+import { ACTION_TYPE_PBFT_GRAPH, COLORS_PBFT_GRAPH, NODES_PBFT_GRAPH, TITLES_PBFT_GRAPH } from "../../../../../Constants";
+import { dummyData } from "../data";
 
 const primaryIndexToPoint = {
     1: 2,
@@ -36,14 +36,10 @@ export const generateConnections = (
 ) => {
     let points = {};
 
-    console.log('DATAAAA', numberOfSteps,
-        xCoords,
-        yCoords,)
-
     let currentData = messageHistory[transactionNumber];
     //If no data, using dummy data to avoid crash
-    if(!currentData){
-        currentData=dummyData[17];
+    if (!currentData) {
+        currentData = dummyData[17];
     }
 
     const { primaryIndex, transactions } =
